Broke model tests out into subdir
Allow the testing and groking of the model and tests.
